+++ front-end/src/app/components/student-account/student-account.component.ts
@@ -1,5 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { StudentService } from '../../shared/services/student.service';
+import { MatDialog } from '@angular/material';
+import { DeleteDialogComponent } from '../../dialogs/delete/delete.dialog.component';
+import { SnackBarService } from '../../shared/services/snackbar.service';
 
 @Component({
   selector: 'app-student-account',
@@ -11,8 +14,13 @@ export class StudentAccountComponent implements OnInit {
 
   draggable = false;
   preferences = null;
+  project = null;
 
-  constructor(private studentService: StudentService) { }
+  constructor(
+    public dialog: MatDialog,
+    private studentService: StudentService,
+    private snackBarService: SnackBarService
+  ) { }
 
   ngOnInit() {
     this.studentService.getPreferences().subscribe(res => {
@@ -21,14 +29,30 @@ export class StudentAccountComponent implements OnInit {
     }, err => {
       console.log('eroare');
     });
+
+    this.studentService.getAssignedProjectForStudent().subscribe(res => {
+      this.project = res;
+      console.log(res);
+    });
   }
 
   removeProject(index: number): void {
-    this.studentService.removePreference(this.preferences[index].id).subscribe(res => {
-      console.log(res);
-      this.preferences.splice(index, 1);
-    }, err => {
-      console.log(err);
+    const dialogRef = this.dialog.open(DeleteDialogComponent, {
+      width: '350px',
+      data: {
+        title: this.preferences[index].project.title
+      }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === 1) {
+        this.studentService.removePreference(this.preferences[index].id).subscribe(res => {
+          this.preferences.splice(index, 1);
+          this.snackBarService.showSnackBar('Preferinta a fost stearsa cu succes!');
+        }, err => {
+          this.snackBarService.showSnackBar('A intervenit o eroare la stergerea preferintei!');
+        });
+      }
     });
   }
